[Bulky Goods] Redirect logged-out user to /auth if following cancel link
If a logged-out user tried to follow the /bulky_cancel link from
an email, for example, they would just be redirected to the bin
days page with no obvious signal that they need to log in to
cancel. Now they are redirected to /auth; if, after logging in,
they are still not authorised to cancel a collection, then we
take them back to bin days.