[Docs] Updates to menu items.
I have updated the names of some of the menu items, removed the link to
NoiseWorks and to the parks/estate pages - both of which are very out of
date. I have removed the individual case studies sub menu, so that,
instead, users can just click directly to the case studies page. I have
removed the features menu - this isn't necessary here.