[Brent] Amend template to deal with '9' representing sacks rather than 9 bins
The BRT - Paid Collection Container Quantity has code 1 matching 1 bin
and continues with 2:2, 3:3 etc so the number code corresponds to the
number of bins.

We have relied on the code for the number of bins.

However, when we reach 9, that corresponds to sacks rather than a number so
if we have 9 bins - it is, in fact, a sack subscription and the template
adapts accordingly.