[Highways England][FMS] Add category checkbox for HE litter
Adds a checkbox when creating/editing a category which can be ticked
to indicate that the category should be considered a litter category
when council are responsible for litter on an HE road.

Maintains the status quo with the current default list. This
hopefully removes risk of a council having no checked categories
and so no litter categories showing up.

https://github.com/mysociety/societyworks/issues/3680