[Bexley] Always use $service->{last} for recent collection checks
There were a couple of places that were trying to get a
recent_collection date using a round_schedule directly, but this doesn't
work because some properties have multiple comma-separated schedules in
the RoundSchedule field.

The fix is to always use $service->{last}, which is calculated based on
all of the rounds mentioned in the RoundSchedule field.