Hide "Report a problem" button in homepage
The "Report a problem" button in the navbar is currently a span element that can't be interacted with in the homepage. That can be confusing for users that want to click the link to report a problem, only to realise it's not link, but an active state to reflect the user is in the homepage where they can start the process to report a problem.