[Zurich][FMS] Adds a filter_description url argument
Can add a filter_description to the url which will only
work when combined with filter_category which allows the
auto-filling of the report detail.

In Zurich's case, it will remove the category choice,
so category can not be changed - also provides a call
hook used by Zurich to allow the reports to be
made in not active categories.

https://github.com/mysociety/societyworks/issues/4517