[offline] On desktop /my/planned don’t cache reports automatically
Desktop users are presumably less likely to find themselves offline,
so it doesn't make sense to cache all reports in the shortlist
automatically. This commit adds a link to the bottom of the report list
on the desktop version of the /my/planned page that allows users to kick
off the caching process manually.

Behaviour on mobile is unchanged.