[UK] Remove broken area check.
`check_report_is_on_cobrand_asset` used shift @_, so $council_area was
always blank. This function is always returning the name of an area if
it matches against the asset layer.

I think the intention was if you clicked on an area that was outside the
boundary, but their responsibility, you'd continue, but outside and
someone else's responsibility, you'd get "go to .com". But what actually
happened was you'd continue on the cobrand site, which worked fine.