Added skip report tools on Around page
This commit should minimise the confusion for assistive device users.
Currently, the way components are presented and the focus order is not
linear; after they have travelled through the map, they are presented
with the 'Can't use the map?' link and then go directly to the
`#key-tools` component sitting at the bottom of the page, skipping the
filters and report list. To respect the page sequence, the user should
go directly to the report filter and then to the report list. The
downside of this approach is that users who want to use the #key-tools
would have to jump through a whole list of reports before being able to
get to the #key-tools. The solution in this commit is to add a link
right after the 'Can't use the map?' that gives the user some context
about the mismatch between the reading sequence and also the opportunity
to skip the #key-tools in case they just want to jump the next element
which would be the report filters.