Accessibility improvements to hint_style for email
Currently, the email's top(email_summary) and bottom(email_footer) parts
don't necessarily pass the contrast test. The text colour uses
body_text_color, and the background colour for the top and bottom
sections is not set, so the hint text is above whatever the
body_background_color is.

We remove email_summary entirely, it doesn't add anything.

The solution on this commit gives hint_style a background that uses
color_white as a variable, and because the text is using the
body_text_color, it is less likely the contrast test will fail. The
scenario where this could happen is when body_text_color is using a
light colour, which, so far, I haven't seen a cobrand doing.