Fixed key-tools problem when small drawer opens
Currently in the all reports page, or view reports by wards
(Oxfordshire), when opening the small drawer the `key-tools` remains
with position fixed. This is because of the CSS hierarchy, which is
overriding the static rule from the `.static` class when the drawer
opens. This commit fixes the hierarchy making the `.static` class to
keep its position.