Update FAQ answer on closed reports.
Update to the answer to the question about reports being marked as
closed or fixed when they are not. This follows a request from user
support pointing out that in some areas it is not possible for
report-makers to update the status of their reports. The FAQ should now
better reflect the reality of the situation.