[Peterborough] Use private land layer for Abandoned vehicles category
This will be shown when people try and make a report on land that
Peterborough aren't responsible for.

I've refactored the asset layers so that there are templates for the PCC
land and PCC leased land, then there's two layers each for
Flytipping/Graffiti/Abandoned vehicles, one for owned and one for
leased.