Removed absolute width from filter fields.
This change is only applied to children of .govuk-fieldset-wrapper.

With this commit they will use all the available space from their parent
container, this should cause less problems when adjusting the sidebar
width in a cobrand. Before the fixed width of 200px caused problems in
some cobrands (National Highways) where the sidebar was not as wide as
in FMS.