[Bexley] Improve Agile API error handling for garden waste
Differentiate between 400/404 error, which are expected when there's no
subscription for a property, and other errors which indicate an
unexpected problem.

When an unexpected problem occurs show the user a helpful message and
remove garden bins from the listing page, as they won't have the correct
data available.

Also adds some debugging info to the superuser debugging panel.