[Bulky] Allow same day amendment by update.
If a backend has no slots left, we cannot use the normal method of
cancelling and creating a new booking (which we still have to do for
date change amendments). Instead, add an amendment moderation entry,
and store the changes in an amendment update for sending through to
the backend (delayed confirmation if further payment is needed).

[SLWP] Include photos in updates, for the amendment images.