[Bexley] Add contract ID lookup for legacy subscriptions
Add BexleyContracts module and make-bexley-contract-db script to handle
UPRN to contract ID lookups for garden waste subscriptions created before
the WasteWorks integration.

First checks the report for a matching DD contract ID, indicating it was
created by WW, and if found uses that, otherwise looks up legacy
contract IDs in this database and attempts to cancel all of those. Also
works for the cancellation script that runs nightly.