[Echo] Store schedule on last instance as well.
Code using this is in general only looking at the schedule on the next
instance; but it is possible, if the round is currently in progress for
the last instance to become the next instance, and then we might still
need to look up the scheduled there (e.g. Merton waste calendar links).