Prevent multiple lookups of user planned reports.
For relevant staff, is_planned_report is being called on the user for
each report in a list in _item.html. This is then performing the same
lookup of that user’s active planned reports over and over. Map pages
are prefetching user_planned_reports in the search, so looking up the
shortlisted_user of a report does not then require another query (and
it is cached for other page use; there are multiple calls on a report
page).